# openapi.yaml
openapi: 3.1.0
info:
  title: Sleep API
  version: '0.1'
paths:
  /api/login:
    post:
      summary: Login (form)
      description: Issues session and CSRF cookies via Set-Cookie. Send subsequent mutating requests with X-CSRF-Token equal to the CSRF cookie value.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginPayload'
      responses:
        '303':
          description: See Other. Form login success redirects to "/".
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/login.json:
    post:
      summary: Login (JSON)
      deprecated: true
      description: Issues session and CSRF cookies via Set-Cookie. JSON variant of login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /api/logout:
    post:
      summary: Logout
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        '204':
          description: No Content. Clears session and CSRF cookies.
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (CSRF)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/session:
    get:
      summary: Session probe
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Authenticated flag
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
  /api/settings/timezone:
    get:
      summary: Get user timezone
      description: Returns the stored IANA timezone name (falls back to APP_TZ when unset).
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimezoneResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Set user timezone
      description: Persist an IANA timezone name for DST-aware duration calculations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimezonePayload'
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        '204':
          description: Updated
        '400':
          description: Invalid timezone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (CSRF)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sleep:
    post:
      description: Creates a sleep session using wake-date semantics. Overlaps are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SleepInput'
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          description: Invalid input (including overlaps)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (CSRF)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sleep/date/{date}:
    get:
      summary: Sleep sessions for a wake date
      description: Returns all sessions that share the same wake date (may be empty).
      parameters:
        - in: path
          name: date
          schema:
            type: string
            format: date
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SleepSession'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sleep/{id}:
    get:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SleepSession'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
    put:
      description: Updates a sleep session. Overlaps are rejected.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SleepInput'
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        '204':
          description: Updated
        '400':
          description: Invalid input (including overlaps)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (CSRF)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        '204':
          description: Deleted or already absent
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (CSRF)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sleep/recent:
    get:
      summary: Recent daily sleep entries
      parameters:
        - in: query
          name: days
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 31
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Up to N recent entries ordered desc by date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SleepListItem'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sleep/range:
    get:
      summary: Sleep sessions within an inclusive date range
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Sessions in range ordered asc by date then wake_time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SleepListItem'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/exercise:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseInput'
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (CSRF)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/exercise/intensity:
    get:
      summary: Exercise intensity by date in range
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Intensities ordered asc by date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DateIntensity'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
      description: Session cookie "__Host-session" (or "session" when COOKIE_SECURE=false)
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Must equal the CSRF cookie value ("__Host-csrf" or "csrf" in dev)
  schemas:
    SleepInput:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Wake date (local). If bed_time > wake_time, bed datetime is treated as previous day.
        bed_time:
          type: string
          format: time
        wake_time:
          type: string
          format: time
        latency_min:
          type: integer
        awakenings:
          type: integer
        quality:
          type: integer
    SleepSession:
      allOf:
        - $ref: '#/components/schemas/SleepInput'
        - type: object
          properties:
            id:
              type: integer
...
    DateIntensity:
      type: object
      properties:
        date:
          type: string
          format: date
        intensity:
          type: string
          enum: [none, light, hard]
...
    LoginPayload:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    TimezonePayload:
      type: object
      properties:
        timezone:
          type: string
          description: IANA timezone name (e.g., Asia/Tokyo)
    TimezoneResponse:
      type: object
      properties:
        timezone:
          type: string
          description: IANA timezone name (e.g., Asia/Tokyo)
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
          nullable: true
...
    BadRequest:
      type: object
      properties:
        code:
          type: string
        message:
          type: string